requirements.txt
discord.py
motor
google-genai
python-dotenv
Pillow